refactor(dependencies): inject service collaborators instead of self-defaulting
Move get_outbound_message_service and get_notification_unit_of_work from
repository_dependency to service_dependency, where objects of their layer
belong.

They were misplaced to avoid an import cycle: NotificationService and
AuthService imported the dependency module to self-default their own
collaborators, so any provider they consumed had to live in whichever module
they imported, and service_dependency already imports NotificationService.

Removing the self-defaulting removes the cycle. Both services now take their
collaborators as required arguments, matching OutboundMessageService, and every
construction site already passed them explicitly, so no caller changes.

LogService and StatsService keep their existing fallbacks; they only reach for
repository providers, so the cycle does not arise there today.
